+++ scripts/lint.sh
@@ -37,7 +37,7 @@ if rustup component add clippy; then
 #  cargo_feature opentelemetry-stackdriver "tls-webpki-roots"
 
   cargo_feature opentelemetry-user-events-logs "default"
-  cargo_feature opentelemetry-user-events-logs "logs_level_enabled"
+  cargo_feature opentelemetry-user-events-logs "spec_unstable_logs_enabled"
 
   cargo_feature opentelemetry-user-events-metrics ""