+++ opentelemetry-datadog/src/exporter/mod.rs
@@ -4,10 +4,12 @@ mod model;
 pub use model::ApiVersion;
 pub use model::Error;
 pub use model::FieldMappingFn;
+use opentelemetry::InstrumentationScope;
 
 use crate::exporter::model::FieldMapping;
 use futures_core::future::BoxFuture;
 use http::{Method, Request, Uri};
+use opentelemetry::trace::TracerProvider as _;
 use opentelemetry::{global, trace::TraceError, KeyValue};
 use opentelemetry_http::{HttpClient, ResponseExt};
 use opentelemetry_sdk::{
@@ -285,16 +287,14 @@ impl DatadogPipelineBuilder {
         let (config, service_name) = self.build_config_and_service_name();
         let exporter = self.build_exporter_with_service_name(service_name)?;
         let mut provider_builder = TracerProvider::builder().with_simple_exporter(exporter);
-        provider_builder = provider_builder.with_config(config);
+        provider_builder = provider_builder.with_resource(config.resource.into_owned());
         let provider = provider_builder.build();
-        let tracer = opentelemetry::trace::TracerProvider::tracer_builder(
-            &provider,
-            "opentelemetry-datadog",
-        )
-        .with_version(env!("CARGO_PKG_VERSION"))
-        .with_schema_url(semcov::SCHEMA_URL)
-        .with_attributes(None)
-        .build();
+        let scope = InstrumentationScope::builder("opentelemetry-datadog")
+            .with_version(env!("CARGO_PKG_VERSION"))
+            .with_schema_url(semcov::SCHEMA_URL)
+            .with_attributes(None)
+            .build();
+        let tracer = provider.tracer_with_scope(scope);
         let _ = global::set_tracer_provider(provider);
         Ok(tracer)
     }
@@ -305,16 +305,14 @@ impl DatadogPipelineBuilder {
         let (config, service_name) = self.build_config_and_service_name();
         let exporter = self.build_exporter_with_service_name(service_name)?;
         let mut provider_builder = TracerProvider::builder().with_batch_exporter(exporter, runtime);
-        provider_builder = provider_builder.with_config(config);
+        provider_builder = provider_builder.with_resource(config.resource.into_owned());
         let provider = provider_builder.build();
-        let tracer = opentelemetry::trace::TracerProvider::tracer_builder(
-            &provider,
-            "opentelemetry-datadog",
-        )
-        .with_version(env!("CARGO_PKG_VERSION"))
-        .with_schema_url(semcov::SCHEMA_URL)
-        .with_attributes(None)
-        .build();
+        let scope = InstrumentationScope::builder("opentelemetry-datadog")
+            .with_version(env!("CARGO_PKG_VERSION"))
+            .with_schema_url(semcov::SCHEMA_URL)
+            .with_attributes(None)
+            .build();
+        let tracer = provider.tracer_with_scope(scope);
         let _ = global::set_tracer_provider(provider);
         Ok(tracer)
     }
@@ -457,6 +455,12 @@ fn mapping_debug(f: &Option<FieldMapping>) -> String {
         "default mapping"
     }
     .to_string()
+}
+
+impl opentelemetry::trace::ExportError for model::Error {
+    fn exporter_name(&self) -> &'static str {
+        "DatadogExporter"
+    }
 }
 
 #[cfg(test)]