+++ opentelemetry-user-events-logs/src/logs/reentrant_logprocessor.rs
@@ -1,8 +1,8 @@
 use std::fmt::Debug;
 
-use opentelemetry::logs::LogResult;
+use opentelemetry_sdk::logs::LogResult;
 
-#[cfg(feature = "logs_level_enabled")]
+#[cfg(feature = "spec_unstable_logs_enabled")]
 use opentelemetry_sdk::export::logs::LogExporter;
 
 use crate::logs::exporter::*;
@@ -29,7 +29,7 @@ impl opentelemetry_sdk::logs::LogProcessor for ReentrantLogProcessor {
     fn emit(
         &self,
         record: &mut opentelemetry_sdk::logs::LogRecord,
-        instrumentation: &opentelemetry::InstrumentationLibrary,
+        instrumentation: &opentelemetry::InstrumentationScope,
     ) {
         _ = self.event_exporter.export_log_data(record, instrumentation);
     }
@@ -46,7 +46,7 @@ impl opentelemetry_sdk::logs::LogProcessor for ReentrantLogProcessor {
         Ok(())
     }
 
-    #[cfg(feature = "logs_level_enabled")]
+    #[cfg(feature = "spec_unstable_logs_enabled")]
     fn event_enabled(
         &self,
         level: opentelemetry::logs::Severity,